Refactored potluck-router tests for DRYer code
